renamed provider.js to dataAccess.js
